Add files via upload
Introduction:¶
In this notebook, we analyze Amazon product reviews to understand the sentiments behind customer feedback. The goal is to determine whether reviews are generally positive, negative, or neutral. Sentiment analysis can help businesses better understand customer perceptions and improve products or services accordingly.

We will use the VADER Sentiment Analyzer, which is a tool in the NLTK library designed to perform sentiment analysis based on text.

Dataset Overview:
The dataset consists of customer reviews and their corresponding ratings for various products sold on Amazon. Each review is accompanied by a numerical rating (Score), which ranges from 1 (lowest) to 5 (highest). Our focus will be on analyzing the text of the reviews to infer the overall sentiment of the customers.